```
The `git status` command tells me that there are some changes that are ready to be committed. The output is quite long, so I have to scroll up to see the changes. The changes are listed under the `Changes to be committed` section.

I have to write a commit message that describes the changes. I can write multiple lines in the commit message. The lines that start with `#` are comments and will be ignored. I have to write a meaningful commit message that explains the Changes